+++ src/app/(afterLogin)/dashboard/page.tsx
