+++ src/app/(afterLogin)/layout.compo.tsx
@@ -1,0 +1,23 @@
+"use client";
+
+// Compo
+import AdminInfoContainer from "@/components/dashboardItems/AdminInfoContainer";
+//
+import React, { useMemo } from "react";
+import { usePathname } from "next/navigation";
+
+const ADMIN_INFOMATION_AVAILABLE_PATHS = [
+  "/dashboard",
+  "/management",
+  "/emergency",
+];
+
+export const AdminInfomation = () => {
+  const pathname = usePathname();
+
+  const isAvailableAdminInfo = useMemo(
+    () => ADMIN_INFOMATION_AVAILABLE_PATHS.some((v) => v === pathname),
+    [pathname],
+  );
+  return isAvailableAdminInfo ? <AdminInfoContainer /> : <></>;
+};
