+++ src/app/(beforeLogin)/auth/page.tsx
@@ -1,5 +1,5 @@
-import Login from "@/components/client/Login.page";
 import React from "react";
+import Login from "@/components/client/Login.page";
 
 const AuthLoginPage = () => {
   return <Login />;
