+++ src/app/layout.tsx
@@ -1,15 +1,15 @@
 // Compo
+
 import ReactCookieProvider from "@/components/ReactCookieProvider";
 import Script from "next/script";
 import GoogleAnalytics from "@/components/GoogleAnalytics";
 
 // Utils
+import React from "react";
 import "@/styles/globals.css";
 import localFont from "next/font/local";
 // Types
 import type { Metadata } from "next";
-import { cookies } from "next/headers";
-import { LayoutSelector } from "./layouts";
 
 const pretendard = localFont({
   src: "./fonts/PretendardVariable.woff2",
@@ -28,9 +28,6 @@ const RootLayout = ({
 }: Readonly<{
   children: React.ReactNode;
 }>) => {
-  const cookieStore = cookies();
-  const isLogined = Boolean(cookieStore.get("authToken"));
-
   return (
     <ReactCookieProvider>
       <html lang="kr">
@@ -47,7 +44,7 @@ const RootLayout = ({
         </head>
         <body className={pretendard.className}>
           <div className="flex flex-col min-h-screen h-auto mainBackGround">
-            <LayoutSelector initialLogin={isLogined}>{children}</LayoutSelector>
+            {children}
           </div>
           {process.env.ANALYSIS || process.env.NODE_ENV === "production" ? (
             <GoogleAnalytics />
