--- src/app/layouts.const.ts
@@ -1,5 +1,0 @@
-export const ADMIN_INFOMATION_AVAILABLE_PATHS = [
-    '/dashboard',
-    '/management',
-    '/emergency'
-]
